feat: implement full-stack order management system with JWT authentication and responsive React UI

+++ order_management/src/main/java/com/example/orderrepository/OrderRepository.java
@@ -6,6 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.example.model.Order;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Integer>{
-
+public interface OrderRepository extends JpaRepository<Order, Long> {
 }
